requirements.txt
franklinwh
google-cloud-secret-manager
functions-framework
pytz
